feat: initial genesys dice
This commit includes the blend file for a tweaked version of the Quest Portal initial dice. It also includes a set of Babylon exports and a draft config file that can be used to roll Genesys dice using the current 1.0.5 dice-box repo. 1.0.5 does not support non d naming hence the dice all prefixed with d. This commit does not include die face mapping.